+++ eslint/eslint-bulk/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/eslint-bulk
 
-This log was last generated on Fri, 21 Nov 2025 16:13:56 GMT and should not be manually modified.
+This log was last generated on Sat, 06 Dec 2025 01:12:28 GMT and should not be manually modified.
+
+## 0.4.8
+Sat, 06 Dec 2025 01:12:28 GMT
+
+_Version update only_
 
 ## 0.4.7
 Fri, 21 Nov 2025 16:13:56 GMT