+++ libraries/package-deps-hash/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @rushstack/package-deps-hash
 
-This log was last generated on Fri, 21 Nov 2025 16:13:56 GMT and should not be manually modified.
+This log was last generated on Fri, 12 Dec 2025 01:12:05 GMT and should not be manually modified.
+
+## 4.6.0
+Fri, 12 Dec 2025 01:12:05 GMT
+
+### Minor changes
+
+- Replace "git ls-tree" with "git ls-files" to improve performance. Identify symbolic links and return them separately in "getDetailedRepoStateAsync". Symbolic links will be omitted from the result returned by "getRepoStateAsync", as they are not "files".
+
+## 4.5.7
+Sat, 06 Dec 2025 01:12:28 GMT
+
+_Version update only_
 
 ## 4.5.6
 Fri, 21 Nov 2025 16:13:56 GMT