+++ webpack/set-webpack-public-path-plugin/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log - @rushstack/set-webpack-public-path-plugin
 
-This log was last generated on Fri, 21 Nov 2025 16:13:56 GMT and should not be manually modified.
+This log was last generated on Thu, 18 Dec 2025 01:13:04 GMT and should not be manually modified.
+
+## 5.2.8
+Thu, 18 Dec 2025 01:13:04 GMT
+
+### Patches
+
+- Fix support for newer webpack typings.
+
+## 5.2.7
+Sat, 06 Dec 2025 01:12:28 GMT
+
+_Version update only_
 
 ## 5.2.6
 Fri, 21 Nov 2025 16:13:56 GMT